+++ matchms/similarity/vector_similarity_functions.py
@@ -4,9 +4,25 @@ import numpy
 
 
 @numba.njit
-def jaccard_similarity_matrix(references, queries):
+def jaccard_similarity_matrix(references: numpy.array, queries: numpy.array) -> numpy.array:
     """Returns matrix of jaccard indices between all-vs-all vectors of references
-    and queries."""
+    and queries.
+
+    Parameters
+    ----------
+    references : numpy.array
+        Reference vectors as 2D numpy array. Expects that vector_i corresponds to
+        references[i, :].
+    queries : numpy.array
+        Query vectors as 2D numpy array. Expects that vector_i corresponds to
+        queries[i, :].
+
+    Returns
+    -------
+    scores : numpy.array
+        Matrix of all-vs-all similarity scores. scores[i, j] will contain the score
+        between the vectors references[i, :] and queries[j, :].
+    """
     size1 = references.shape[0]
     size2 = queries.shape[0]
     scores = numpy.zeros((size1, size2))
@@ -17,9 +33,25 @@ def jaccard_similarity_matrix(references, queries):
 
 
 @numba.njit
-def dice_similarity_matrix(references, queries):
+def dice_similarity_matrix(references: numpy.array, queries: numpy.array) -> numpy.array:
     """Returns matrix of dice similarity scores between all-vs-all vectors of references
-    and queries."""
+    and queries.
+
+    Parameters
+    ----------
+    references : numpy.array
+        Reference vectors as 2D numpy array. Expects that vector_i corresponds to
+        references[i, :].
+    queries : numpy.array
+        Query vectors as 2D numpy array. Expects that vector_i corresponds to
+        queries[i, :].
+
+    Returns
+    -------
+    scores : numpy.array
+        Matrix of all-vs-all similarity scores. scores[i, j] will contain the score
+        between the vectors references[i, :] and queries[j, :].
+    """
     size1 = references.shape[0]
     size2 = queries.shape[0]
     scores = numpy.zeros((size1, size2))
@@ -30,9 +62,25 @@ def dice_similarity_matrix(references, queries):
 
 
 @numba.njit
-def cosine_similarity_matrix(references, queries):
+def cosine_similarity_matrix(references: numpy.array, queries: numpy.array) -> numpy.array:
     """Returns matrix of cosine similarity scores between all-vs-all vectors of
-    references and queries."""
+    references and queries.
+
+    Parameters
+    ----------
+    references : numpy.array
+        Reference vectors as 2D numpy array. Expects that vector_i corresponds to
+        references[i, :].
+    queries : numpy.array
+        Query vectors as 2D numpy array. Expects that vector_i corresponds to
+        queries[i, :].
+
+    Returns
+    -------
+    scores : numpy.array
+        Matrix of all-vs-all similarity scores. scores[i, j] will contain the score
+        between the vectors references[i, :] and queries[j, :].
+    """
     size1 = references.shape[0]
     size2 = queries.shape[0]
     scores = numpy.zeros((size1, size2))
@@ -43,7 +91,7 @@ def cosine_similarity_matrix(references, queries):
 
 
 @numba.njit
-def jaccard_index(u, v):
+def jaccard_index(u: numpy.array, v: numpy.array) -> numpy.float64:
     r"""Computes the Jaccard-index (or Jaccard similarity coefficient) of two boolean
     1-D arrays.
     The Jaccard index between 1-D boolean arrays `u` and `v`,
@@ -54,26 +102,26 @@ def jaccard_index(u, v):
 
     Args:
     ----
-    u : (N,) array_like, bool
+    u : (N,) numpy.array, bool
         Input array.
-    v : (N,) array_like, bool
+    v : (N,) numpy.array, bool
         Input array.
 
     Returns
     -------
-    jaccard_similarity : double
+    jaccard_similarity : float
         The Jaccard similarity coefficient between vectors `u` and `v`.
     """
     u_or_v = numpy.bitwise_or(u != 0, v != 0)
     u_and_v = numpy.bitwise_and(u != 0, v != 0)
     jaccard_score = 0
     if u_or_v.sum() != 0:
-        jaccard_score = numpy.double(u_and_v.sum()) / numpy.double(u_or_v.sum())
+        jaccard_score = numpy.float64(u_and_v.sum()) / numpy.float64(u_or_v.sum())
     return jaccard_score
 
 
 @numba.njit
-def dice_similarity(u, v):
+def dice_similarity(u: numpy.array, v: numpy.array) -> numpy.float64:
     r"""Computes the Dice similarity coefficient (DSC) between two boolean 1-D arrays.
 
     The Dice similarity coefficient between `u` and `v`, is
@@ -83,38 +131,38 @@ def dice_similarity(u, v):
 
     Args:
     ----
-    u : (N,) array_like, bool
+    u : (N,) numpy.array, bool
         Input array.
-    v : (N,) array_like, bool
+    v : (N,) numpy.array, bool
         Input array.
 
     Returns
     -------
-    dice_similarity : double
+    dice_similarity : float
         The Dice similarity coefficient between 1-D arrays `u` and `v`.
     """
     u_and_v = numpy.bitwise_and(u != 0, v != 0)
     u_abs_and_v_abs = numpy.abs(u).sum() + numpy.abs(v).sum()
     dice_score = 0
     if u_abs_and_v_abs != 0:
-        dice_score = 2.0 * numpy.double(u_and_v.sum()) / u_abs_and_v_abs
+        dice_score = 2.0 * numpy.float64(u_and_v.sum()) / numpy.float64(u_abs_and_v_abs)
     return dice_score
 
 
 @numba.njit
-def cosine_similarity(u, v):
+def cosine_similarity(u: numpy.array, v: numpy.array) -> numpy.float64:
     """Calculate cosine similarity score.
 
     Args:
     ----
-    u : numpy array, float
+    u : numpy.array
         Input vector.
-    v : numpy array, float
+    v : numpy.array
         Input vector.
 
     Returns
     -------
-    cosine_similarity : double
+    cosine_similarity : float
         The Jaccard similarity coefficient between vectors `u` and `v`.
     """
     assert u.shape[0] == v.shape[0], "Input vector must have same shape."
@@ -128,4 +176,4 @@ def cosine_similarity(u, v):
     cosine_score = 0
     if uu != 0 and vv != 0:
         cosine_score = uv / numpy.sqrt(uu * vv)
-    return cosine_score
+    return numpy.float64(cosine_score)